requirements.txt
# HTTP requests and session management
requests>=2.31.0

# HTML parsing
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Terminal formatting for tables(if u want to)
tabulate>=0.9.0

# Environment variable management (for credentials)
python-dotenv>=1.0.0
